(MAINT) Ensure temp file is xplat for gem testing
Prior to this commit the temp file was /tmp/out.tmp, a directory which is
unlikely to exist on windows systems. This commit updates the logic and
removes the hardcoding for the path, utilitizing the TEMP environment
variable on windows or /tmp on *nix.

This prevents the tasks from exploding on directory not found when run
against a windows machine.