(MAINT) Cleanup repo parsing
Prior to this commit the repository parsing for a gem line worked only
on non-Windows systems as the line being passed was specific to the
escape sequences there.

This commit updates the parse to strip any additional whitespace and to
drop any single quotes around the repo URI as git clone will explode if
the URI is wrapped in single quotes.

This commit also updates the alternative parser to use the namespace from
the configuration since it was erroneously nil otherwise.