Removal of old obsolete gems
Prior to this commit we were using old gems that are no longer used. We don't use Travis anymore and also the changelog generator. As they are old dependencies they are brining in a version of activesupport that has a CVE.

Removing the gems will stop us bringing in old dependencies we no longer need.